serverless: refactor structure to match associated untyped project
