+++ types/serverless/plugins/aws/provider/awsProvider.d.ts
@@ -1,0 +1,12 @@
+import Serverless = require("../../../index");
+
+declare class Aws {
+    constructor(serverless: Serverless, options: Serverless.Options)
+
+    getProviderName(): string;
+    getRegion(): string;
+    getServerlessDeploymentBucketName(): string;
+    getStage(): string;
+}
+
+export = Aws;
